backend/requirements.txt
# Core Dependencies
Flask==3.1.1
Werkzeug==3.1.3
python-dotenv==1.1.0

# Flask Extensions
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1
alembic==1.16.1

# LangChain Ecosystem
langchain==0.3.25
langchain-core==0.3.62
langgraph-checkpoint-sqlite
langgraph-core
flask-sock==0.7.0
langgraph==0.4.7
langchain-google-vertexai==2.0.24
langchain-google-genai==2.1.4
langchain-community==0.3.24
langsmith==0.3.42

# Google Cloud Services
google-cloud-aiplatform==1.94.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-cloud-storage==2.19.0
google-cloud-documentai==3.5.0
google-cloud-firestore==2.20.2
firebase-admin==6.8.0

# Database
SQLAlchemy==2.0.41

# Data Processing
numpy==2.2.6
pandas==2.2.3
PyPDF2==3.0.1

# Utilities
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic-extra-types==2.10.4
requests==2.32.3
protobuf==6.31.0
grpcio==1.73.0rc1

# Testing
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.1
pytest-asyncio==1.0.0

# Runtime Dependencies
python-engineio==4.12.1
python-socketio==5.13.0
pyjwt[crypto]==2.10.1
